requirements.txt
asgiref==3.4.1
backcall==0.2.0
decorator==5.0.9
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.2
numpy==1.21.2
pandas==1.3.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
psycopg2==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
traitlets==5.1.0
wcwidth==0.2.5
